Experimental >2 billion event support

+++ PianoFromAbove/GameState.h
@@ -155,8 +155,8 @@ public:
 };
 
 typedef struct {
-    int idx;
-    int sister_idx;
+    unsigned idx;
+    unsigned sister_idx;
 } thread_work_t;
 
 class MainScreen : public GameState
@@ -250,7 +250,7 @@ private:
 
     // Playback
     State m_eGameMode;
-    int m_iStartPos, m_iEndPos; // Postions of the start and end events that occur in the current window
+    long long m_iStartPos, m_iEndPos; // Postions of the start and end events that occur in the current window
     long long m_llStartTime, m_llTimeSpan;  // Times of the start and end events of the current window
     int m_iStartTick; // Tick that corresponds with m_llStartTime. Used to help with beat and metronome detection
     vector<int> m_vState[128];  // The notes that are on at time m_llStartTime.
